chore: 更新类型文件
