+++ src/components/Layers/BaseLayers/HeatmapLayer/types.ts
@@ -1,7 +1,7 @@
 import type { HeatmapLayer, HeatmapLayerOptions } from '@antv/l7-composite-layers';
-import type { LayerCommonProps } from '../../../../types/common';
+import type { LayerCommonProps } from '../../../../types';
 
 /**
  * 组件类型定义
  */
-export interface HeatmapLayerProps extends HeatmapLayerOptions, LayerCommonProps<HeatmapLayer> {}
+export interface HeatmapLayerProps extends Omit<HeatmapLayerOptions, 'onClick'>, LayerCommonProps<HeatmapLayer> {}
