+++ src/components/Layers/BaseLayers/LineLayer/types.ts
@@ -1,7 +1,7 @@
 import type { LineLayer, LineLayerOptions } from '@antv/l7-composite-layers';
-import type { LayerCommonProps } from '../../../../types/common';
+import type { LayerCommonProps } from '../../../../types';
 
 /**
  * 组件类型定义
  */
-export interface LineLayerProps extends LineLayerOptions, LayerCommonProps<LineLayer> {}
+export interface LineLayerProps extends Omit<LineLayerOptions, 'onClick'>, LayerCommonProps<LineLayer> {}
