+++ src/components/Layers/BaseLayers/PointLayer/types.ts
@@ -1,7 +1,7 @@
 import type { PointLayer, PointLayerOptions } from '@antv/l7-composite-layers';
-import type { LayerCommonProps } from '../../../../types/common';
+import type { LayerCommonProps } from '../../../../types';
 
 /**
  * 组件类型定义
  */
-export interface PointLayerProps extends PointLayerOptions, LayerCommonProps<PointLayer> {}
+export interface PointLayerProps extends Omit<PointLayerOptions, 'onClick'>, LayerCommonProps<PointLayer> {}
