+++ src/components/Layers/BaseLayers/PolygonLayer/types.ts
@@ -1,7 +1,7 @@
 import type { PolygonLayer, PolygonLayerOptions } from '@antv/l7-composite-layers';
-import type { LayerCommonProps } from '../../../../types/common';
+import type { LayerCommonProps } from '../../../../types';
 
 /**
  * 组件类型定义
  */
-export interface PolygonLayerProps extends PolygonLayerOptions, LayerCommonProps<PolygonLayer> {}
+export interface PolygonLayerProps extends Omit<PolygonLayerOptions, 'onClick'>, LayerCommonProps<PolygonLayer> {}
