+++ src/types/common.ts
@@ -6,8 +6,3 @@ export type CommonProps = {
   /** 行内样式 */
   style?: CSSProperties;
 };
-
-export interface LayerCommonProps<T> {
-  /** 图层初始化完成后回调，用于获取 layer 对象   */
-  onCreated?: (layer: T) => void;
-}
